+++ NSU_FESTIVAL/src/main/java/com/example/nsu_festival/domain/booth/entity/Booth.java
@@ -28,7 +28,7 @@ public class Booth {
     @NonNull
     private String content;
 
-    private Long countLike;
+    private int countLike;
 
     private String area;
